+++ ExperimentalEnemyInteractions/Generics/Networking.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Text;
 using LethalNetworkAPI;
 using HarmonyLib;
 using BepInEx.Logging;