+++ ExperimentalEnemyInteractions/Script.cs
@@ -9,8 +9,6 @@ using BepInEx.Configuration;
 using System.Text;
 using BepInEx.Bootstrap;
 using NaturalSelection.Compatibility;
-using System.Reflection;
-using UnityEngine;
 //using NetcodePatcher;
 
 namespace NaturalSelection;